requirements.txt
Django==5.2.1
python-dotenv==1.0.0
langchain-groq==0.2.1
langchain-core==0.3.28
torch==2.5.1
transformers==4.47.1
